+++ docs/release-notes.md
@@ -5,6 +5,12 @@ hide:
 
 # Release Notes
 
+## 0.17.3
+
+### Fixed
+
+- Lazy ManyToMany fields.
+
 ## 0.17.2
 
 ### Added
@@ -14,7 +20,7 @@ hide:
 ### Fixed
 
 - `update` and `delete` methods of QuerySet did ignore or clauses.
-- Fixed ManyToMany fields not able to use their Owner Model as target.
+- Fixed ManyToMany fields not able to use their owner Model as target.
 - Fixed makemigration failing with foreignkeys when using model_apps.
 
 ## 0.17.1